fix datatype mismatch when selecting glyphs in gamera gui

+++ gamera/gui/gamera_display.py
@@ -469,8 +469,8 @@ class ImageDisplay(wx.ScrolledWindow, util.CallbackObject):
          dc.SetPen(wx.TRANSPARENT_PEN)
          brush = wx.BLUE_BRUSH
          dc.SetBrush(brush)
-         self.block_w = block_w = max(min((w / 2 - 1)/scaling,8), 4)
-         self.block_h = block_h = max(min((h / 2 - 1)/scaling,8), 4)
+         self.block_w = block_w = max(min(int((w / 2 - 1) / scaling), 8), 4)
+         self.block_h = block_h = max(min(int((h / 2 - 1) / scaling), 8), 4)
          dc.DrawRectangle(x + 1, y + 1, block_w, block_h)
          dc.DrawRectangle(x2 - block_w - 1, y + 1, block_w, block_h)
          dc.DrawRectangle(x + 1, y2 - block_h - 1, block_w, block_h)
